compiling modified hypervisor completed example
