+++ Cargo.toml
@@ -16,7 +16,8 @@ members = [
 qx86 = { git = "https://github.com/qtumproject/qx86-rs", branch = "master"}
 struct_deser = "0.1.1"
 struct_deser-derive = "0.1.0"
-neutron-star-constants = { git = "https://github.com/qtumproject/neutronstar-constants", branch = "master"}
+#neutron-star-constants = { git = "https://github.com/qtumproject/neutronstar-constants", branch = "master"}
+neutron-star-constants = { path = "../neutron-star-constants"}
 ring = "0.16.9"
 num = "0.2.0"
 num-derive = "0.3.0"
